+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 - Fixed an issue in `apps:create` where App Store ID was always prompted for even when unnecessary.
 - Add `functions:lifecycle:list` and `functions:lifecycle:run` commands to view and run
   lifecycle hooks in isolation.
+- Update `firebase init` to accept multiple features from command line